Scripts: add script to split joined subtitle, extend index TSV file format
The new cut-edited-subtitles script can take a subtitle for the entire
fanedit and split it into small chapter/run files.

The code in cut-source-subtitles was moved to a module so that the code
can be shared.

Finally fix a bug in the offsetting script: it would still expect
numeric run ids, whereas those are now only used when we've got no
choice.