NL: Improvements, fix errors
Sometimes a couple of subtitles read better when merged.

Then there are also places where copy/paste errors occurred or where the
timings are off.

Finally eradicate most uses of italics. While it is used in the English
template to indicate a voice-over, there are few scenes where one could
be confused about who is talking. In this case the italics are just
annoying when used for a long while.