DE: Add subtitle for Update 15
These have been pasted over Eric's template files so that they're sure
to be in sync.

However, German is a wordy language, and so it will still take some
review and rewriting to get some titles to fit the allotted time.

The first chapter has seen the most work so far. Things which clearly
need review have been marked with "!!".