Scripts: rework offset-titles to work in-place, forwards and backwards
The script no longer uses Subtitle Edit for its offsetting work, instead
SRT files are written using nothing but Powershell.

The new code supports updating the offsets in-place (although it can
still write to a different folder if needed), and it has a "-Reverse"
switch to restore a particular file to the time base of the original
films.

The script also has some protection built-in against offsetting twice.
SRT files will be scanned to see if they match the intervals defined in
the index, and if one or more files are far outliers then the script
will ask for confirmation.