Index: rework source start and end offsets
Splitting fanedit subtitles made it clear that Gandalf's last line was
not detected as part of c03r10, which means the end boundary was
incorrect.

Similar problems occurred in c03r03 and c03r04, but this time it might
have been because c03r04's start value was too early.