DE: Rework subtitles to be 44 chars max per single line
Previously there were lines with 49 characters, which is too long.

Some single lines could just be split in two, but others either needed
to be rewritten to use shorter words, or the subtitles were split so
that two short-lived subtitles can at least show all words (in that case
care was taken to sync to the spoken dialog if possible).