FR: Rework subtitle for timing and small errors, credits
The basic translation template is pretty good for most of the film, but
there are a few parts where a French translation simply needs more
characters on screen to get the same meaning across.

That changes the Characters Per Second statistics, and as a result there
have been a sizable number of places where lead-in and lead-out times
have been tweaked, subtitles split or even merged in different ways.

Very rarely has a word in the text been removed or replaced to lower the
CPS, and there has been one place where a complete rewording was
necessary.

Finally a very small number of mistakes have been fixed. Most are truly
minor, like some text being on the wrong side of a split subtitle.
Others are about missing text, or text being pasted at the wrong time.

The credits have been modified to include JBert on the Synchronisation
line.