Index, scripts: make sure that versioning info is processed
The "versioning" info at the end of separate SRTs was not picked up.

To do it the index needed to be changed, and then a bug in the subtitle
parser would lose the last subtitle in an SRT file if there was no empty
line at the end of the file.